Cleanup on directorupdate
